+++ cubi_sak/snappy/pull_sheet.py
@@ -8,20 +8,27 @@ More Information
 - `BiomedSheet Documentation <https://biomedsheets.readthedocs.io/en/master/>`__.
 """
 
-import os
 import argparse
+import difflib
+import os
+import shutil
+import tempfile
 from uuid import UUID
 import re
 import sys
 import typing
 
+import icdiff
 from logzero import logger
 import requests
+from termcolor import colored
 
 from .. import exceptions
 
 
 #: The URL template to use.
+from ..common import get_terminal_columns
+
 URL_TPL = "%(sodar_url)s/samplesheets/api/remote/get/%(project_uuid)s/%(api_key)s"
 
 #: Template for the to-be-generated file.
@@ -82,6 +89,25 @@ def setup_argparse(parser: argparse.ArgumentParser) -> None:
     )
 
     parser.add_argument(
+        "--dry-run",
+        default=False,
+        action="store_true",
+        help="Perform a dry run, i.e., don't change anything only display change, implies '--show-diff'.",
+    )
+    parser.add_argument(
+        "--show-diff",
+        default=False,
+        action="store_true",
+        help="Show change when creating/updating sample sheets.",
+    )
+    parser.add_argument(
+        "--show-diff-side-by-side",
+        default=False,
+        action="store_true",
+        help="Show diff side by side instead of unified.",
+    )
+
+    parser.add_argument(
         "--library-types", help="Library type(s) to use, comma-separated, default is to use all."
     )
 
@@ -125,7 +151,7 @@ def check_args(args) -> int:
         if not args.allow_overwrite:
             logger.error(
                 "The output path %s already exists but --allow-overwrite not given.",
-                args.output_tsv,
+                args.output_tsv.name,
             )
             any_error = True
         else:
@@ -144,16 +170,8 @@ def check_args(args) -> int:
     return int(any_error)
 
 
-def run(
-    args, _parser: argparse.ArgumentParser, _subparser: argparse.ArgumentParser
-) -> typing.Optional[int]:
-    """Run ``cubi-sak snappy pull-sheet``."""
-    res = check_args(args)
-    if res:  # pragma: nocover
-        return res
-
-    logger.info("Starting to pull sheet...")
-    logger.info("  Args: %s", args)
+def write_sheet(args, sheet_file) -> typing.Optional[int]:
+    """Write sheet to ``sheet_file``."""
 
     # Query investigation JSON from API.
     url = URL_TPL % {
@@ -239,10 +257,10 @@ def run(
 
     # Generate the resulting sample sheet.
     try:
-        args.output_tsv.truncate()
+        sheet_file.truncate()
     except OSError:  # pragma: nocover
         logger.debug("Could not truncate output TSV (stdout/stderr)? Continuing...")
-    print("\n".join(HEADER_TPL) % vars(args), file=args.output_tsv)
+    print("\n".join(HEADER_TPL) % vars(args), file=sheet_file)
     for source, info in study_map.items():
         if source not in assay_map:  # pragma: nocover
             logger.info("source %s does not have an assay.", source)
@@ -301,6 +319,87 @@ def run(
             seq_platform,
             library_kit,
         ]
-        print("\t".join(row), file=args.output_tsv)
+        print("\t".join(row), file=sheet_file)
+
+    logger.debug("Done writing temporary file.")
+    return None
+
+
+def run(
+    args, _parser: argparse.ArgumentParser, _subparser: argparse.ArgumentParser
+) -> typing.Optional[int]:
+    """Run ``cubi-sak snappy pull-sheet``."""
+    res: typing.Optional[int] = check_args(args)
+    if res:  # pragma: nocover
+        return res
+
+    logger.info("Starting to pull sheet...")
+    logger.info("  Args: %s", args)
+
+    with tempfile.NamedTemporaryFile(mode="w+t") as sheet_file:
+        # Write sheet to temporary file.
+        res = write_sheet(args, sheet_file)
+        if res:  # pragma: nocover
+            return res
+
+        # Compare sheet with output if exists and --show-diff given.
+        if args.show_diff:
+            if os.path.exists(args.output_tsv.name):
+                with open(args.output_tsv.name, "rt") as inputf:
+                    old_lines = inputf.read().splitlines(keepends=False)
+            else:
+                old_lines = []
+            sheet_file.seek(0)
+            new_lines = sheet_file.read().splitlines(keepends=False)
+
+            if not args.show_diff_side_by_side:
+                lines = difflib.unified_diff(
+                    old_lines, new_lines, fromfile=args.output_tsv.name, tofile=args.output_tsv.name
+                )
+                for line in lines:
+                    line = line[:-1]
+                    if line.startswith(("+++", "---")):
+                        print(colored(line, color="white", attrs=("bold",)), file=sys.stdout)
+                    elif line.startswith("@@"):
+                        print(colored(line, color="cyan", attrs=("bold",)), file=sys.stdout)
+                    elif line.startswith("+"):
+                        print(colored(line, color="green", attrs=("bold",)), file=sys.stdout)
+                    elif line.startswith("-"):
+                        print(colored(line, color="red", attrs=("bold",)), file=sys.stdout)
+                    else:
+                        print(line, file=sys.stdout)
+            else:
+                cd = icdiff.ConsoleDiff(cols=get_terminal_columns(), line_numbers=True)
+                lines = cd.make_table(
+                    old_lines,
+                    new_lines,
+                    fromdesc=args.output_tsv.name,
+                    todesc=args.output_tsv.name,
+                    context=True,
+                    numlines=3,
+                )
+                for line in lines:
+                    line = "%s\n" % line
+                    if hasattr(sys.stdout, "buffer"):
+                        sys.stdout.buffer.write(line.encode("utf-8"))
+                    else:
+                        sys.stdout.write(line)
+
+            sys.stdout.flush()
+            if not lines:
+                logger.info("File %s not changed, no diff...", args.output_tsv.name)
+
+        # Write to output file if not --dry-run is given
+        if hasattr(args.output_tsv, "name") and args.dry_run:
+            logger.warn("Not changing %s as we are in --dry-run mode", args.output_tsv.name)
+        else:
+            if hasattr(args.output_tsv, "name"):
+                action = "Overwriting" if os.path.exists(args.output_tsv.name) else "Creating"
+                logger.info("%s %s", action, args.output_tsv.name)
+            sheet_file.seek(0)
+            if hasattr(args.output_tsv, "name"):
+                args.output_tsv.seek(0)
+                args.output_tsv.truncate()
+            shutil.copyfileobj(sheet_file, args.output_tsv)
 
     return None